cloud-init: use vcli to handle user scripts
Instead of creating a shell script in util.runcmd(), create a vcli script.
Additionally, scripts-per-{once,bootinstance} will now also be run as
vcli scripts in util.runparts().

There appears to be a bug in util.runparts() with respect to exe_prefix.
This appears to be list of iterative prefixes but it isn't use that way.
Ignore for now since we don't care.

Enhancement: VRVDR-14043