Fix SSH config and host key generation
The vyatta-update-ssh.pl script which generates the SSH config and host
keys queries the config, which means it must be run in the context of a
config session.

This fix runs the script in a config session, which resolves the issue
where a reboot was required between cloud-init completing and the SSH
server being usable.

VRVDR-48124